Added tnoremap to my keymap functions
This was to change the obnoxious <C-\><C-n> mapping on nvim terminal
That mapping changes you from terminal mode into normal mode, but it's
hideous and weird. But changing it to <Esc> breaks telescope and fzf so
I'm using <C-w>n as a replacement. Still weird but more familiar at
least